Create open class for scheduler
